Grouped schema retrieval methods into a utility class
