+++ dockstore-webservice/pom.xml
@@ -523,7 +523,7 @@
         <dependency>
             <groupId>com.fasterxml.jackson.dataformat</groupId>
             <artifactId>jackson-dataformat-yaml</artifactId>
-            <version>2.7.8</version>
+            <version>${jackson.version}</version>
         </dependency>
         <dependency>
             <groupId>com.github.jsonld-java</groupId>
