+++ dockstore-webservice/src/test/resources/noSchema.json
@@ -1,0 +1,1 @@
+[]
